refactor: Dao 구현체가 아닌 인터페이스에 의존하도록 변경
- 콘솔 어플리케이션도 저장 기능 구현 예정
- 하나의 서비스가 여러 저장 방식을 수행 가능하게 하기 위한 추상
- 스프링에서 생성자에 인터페이스 타입으로 정의했는데 어떻게 구현체를 찾아서 매핑하는걸까?
