+++ src/test/java/racingcar/dao/JdbcPlayRecordsDaoTest.java
@@ -6,13 +6,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import racingcar.repository.dao.PlayRecordsDao;
+import racingcar.repository.dao.JdbcPlayRecordsDao;
 
 @SpringBootTest
-class PlayRecordsDaoTest {
+class JdbcPlayRecordsDaoTest {
 
     @Autowired
-    private PlayRecordsDao playRecordsDao;
+    private JdbcPlayRecordsDao playRecordsDao;
 
     @DisplayName("DB: 게임 이력 저장 테스트")
     @Test
